% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateConfidenceLevel1.R
\name{AnnotateConfidenceLevel1}
\alias{AnnotateConfidenceLevel1}
\title{AnnotateConfidenceLevel1}
\usage{
AnnotateConfidenceLevel1(
  experimental.values,
  theoretical.values,
  mz.flexibility,
  rt.flexibility
)
}
\arguments{
\item{experimental.values}{A dataframe of unknown experimental values, created by the user.
It is important to note that it must be in the following format, including capitalization and class:
mz (numeric), rt (numeric, in seconds), column (numeric), z (numeric), MS2 (in concatenated format, character).}

\item{theoretical.values}{TODO A dataframe of known, theoretical values, i.e, the Ingalls Standards list.
Like its partner the experimental.values dataframe, the theoretical.values dataframe columns need to be in
the following exact format: compound (character), mz (numeric), rt (numeric, in seconds),
column (all caps, character), z (numeric), OPTIONAL filename (character), and MS2 (concatenated format, character).}

\item{mz.flexibility}{Flexibility for m/z matching between experimental and theoretical values. Usually defined as 0.02.}

\item{rt.flexibility}{Flexibility for retention time matching between experimental and theoretical values. Usually defined as ~ 15-30 seconds.}
}
\value{
A complete dataframe, annotated for Confidence Level 1.
}
\description{
Compare a dataframe of experimental values with a dataframe of all theoretical values from the Ingalls
Lab Standards sheet.
}
\examples{
library(phobos)
# Isolate experimental data. Remember that phobos REQUIRES all experimental dataframes to be in the following format:
# - "MassFeature": Your unique mass feature, character.
# - "mz": The mz value, numeric.
# - "rt": The retention time, in seconds, numeric.
# - "column": Column the mass feature was run on, character.
# - "z": The ion mode, numeric.
# - "MS2": MS2 data for those compounds that have it, character, in concatenated format.

example_dir <- system.file("example_data", package = "phobos")
example_data <- list.files(example_dir, full.names = TRUE)
experimental.values <- read.csv(grep("Example_Experimental", example_data, value = TRUE))
theoretical.values <- read.csv(grep("Theoretical_Data", example_data, value = TRUE))
example_confidenceL1 <- AnnotateConfidenceLevel1(experimental.values = experimental.values, theoretical.values = theoretical.values,
mz.flexibility = 0.02, rt.flexibility = 30)
}
