% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateConfidenceLevel4.R
\name{AnnotateConfidenceLevel4}
\alias{AnnotateConfidenceLevel4}
\title{Annotate Confidence Level 4: Everything else!}
\usage{
AnnotateConfidenceLevel4(Confidence.Level.3)
}
\arguments{
\item{Confidence.Level.3}{A dataframe annotated for Confidence Levels 1, 2, & 3. Level 3 must have
received annotation from both MassBank and KEGG! Please refer to the README for details.}
}
\value{
A fully annotated dataframe, annotated for Confidence Levels 1, 2, 3, & 4.
}
\description{
Annotate Confidence Level 4: Everything else!
}
\examples{
#' library(phobos)
# Load experimental data that has already been annotated for Confidence Levels 1, 2, & 3.
Confidence.Level.3_KEGG <- read.csv("data/Example_ConfidenceLevel3_KEGG.csv")

example_confidenceL4 <- AnnotateConfidenceLevel4(Confidence.Level.3 = Confidence.Level.3_KEGG)
}
