% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateKEGGConfidenceLevel3.R
\name{AnnotateKEGGConfidenceLevel3}
\alias{AnnotateKEGGConfidenceLevel3}
\title{Annotate Confidence Level 3: KEGG}
\usage{
AnnotateKEGGConfidenceLevel3(
  Confidence.Level.3.MoNA,
  KEGG.Data,
  mz.flexibility,
  ppm.tolerance
)
}
\arguments{
\item{Confidence.Level.3.MoNA}{A dataframe, already annotated for confidence levels 1, 2 (from Massbank) and 3 (from Massbank).
Please refer to the README or documentation for the AnnotateConfidenceLevel1, AnnotateMoNAConfidenceLevel2, and AnnotateMoNAConfidenceLevel3 functions for details.}

\item{KEGG.Data}{Compound information from the KEGG database, scraped and downloaded. This csv is available on the Ingalls Shared Drive in the MARS_Project folder, titled KEGG_Data.csv}

\item{mz.flexibility}{Flexibility for m/z matching between experimental and theoretical values. Usually defined as 0.02.}

\item{ppm.tolerance}{Parts per million tolerance for matching between experimental and KEGG data. Usually defined as 15.}
}
\value{
A dataframe annotated for Confidence Levels 1, 2, and 3 using the Ingalls standards, MassBank of North America, and KEGG.
}
\description{
Annotate Confidence Level 3: KEGG
}
\examples{
library(phobos)
# Load experimental data that has already been annotated for Confidence Levels 1, 2 & 3
# using the AnnotateConfidenceLevel1, AnnotateMoNAConfidenceLevel2 and AnnotateMoNAConfidenceLevel3 functions.

Confidence.Level.3.MoNA <- read.csv("example_data/Example_ConfidenceLevel3MoNA.csv")
KEGG.Data <- read.csv("your/file/path/here/KEGG_Data.csv")

Example_KEGG_ConfidenceLevel3 <- AnnotateKEGGConfidenceLevel3(Confidence.Level.3.MoNA = Confidence.Level.3.MoNA,
KEGG.Data <- KEGG_Data, mz.flexibility = 0.02, ppm.tolerance = 15)
}
