% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateMoNAConfidenceLevel3.R
\name{AnnotateMoNAConfidenceLevel3}
\alias{AnnotateMoNAConfidenceLevel3}
\title{Annotate Confidence Level 3: MoNA}
\usage{
AnnotateMoNAConfidenceLevel3(
  Confidence.Level.2,
  MassBank.Neg,
  MassBank.Pos,
  mz.flexibility
)
}
\arguments{
\item{Confidence.Level.2}{A dataframe of experimental values, already annotated for Confidence Levels 1 & 2.
Please refer to the README or documentation for the AnnotateConfidenceLevel1 and AnnotateMoNAConfidenceLevel2 functions for details.}

\item{MassBank.Neg}{Spectra in negative ion mode from MassBank of North America, scraped and downloaded. A modified example is included in this package,
but the complete csv is available on the Ingalls Shared Drive in the MARS_Project folder, titled NEG_Spectra.csv.}

\item{MassBank.Pos}{Spectra in positive ion mode from MassBank of North America, scraped and downloaded. A modified example is included in this package,
but the complete csv is available on the Ingalls Shared Drive in the MARS_Project folder, titled POS_Spectra.csv.}

\item{mz.flexibility}{Flexibility for m/z matching between experimental and theoretical values. Usually defined as 0.02.}
}
\value{
A dataframe annotated for Confidence Level 3 by referencing MassBank of North America.
}
\description{
Annotate Confidence Level 3: MoNA
}
\examples{
library(phobos)
# Load experimental data that has already been annotated for Confidence Levels 1 & 2
# using the AnnotateConfidenceLevel1 and AnnotateMoNAConfidenceLevel2 function2.
# Remember that dataframe format is important: please refer to README or the documentation for AnnotateConfidenceLevel1() for details.

Confidence.Level.2 <- read.csv("data/Example_ConfidenceLevel2.csv")
MassBank.Neg <- read.csv("data/NEG_Spectra.csv")
MassBank.Pos <- read.csv("data/POS_Spectra.csv")

Example_ConfidenceLevel3 <- AnnotateMoNaConfidenceLevel3(Confidence.Level.2 = Confidence.Level.2,
MassBank.Neg = MassBank.Neg, MassBank.Pos = MassBank.Pos, mz.flexibility = 0.02)
}
