% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CalculateSimilarityScore}
\alias{CalculateSimilarityScore}
\title{Calculate value similarity score.}
\usage{
CalculateSimilarityScore(experimental.value, theoretical.value, flexibility)
}
\arguments{
\item{experimental.value}{The experimental value to be compared to theoretical.}

\item{theoretical.value}{The theoretical value for comparison.}

\item{flexibility}{User-defined search tolerance, also used as the standard deviation value.}
}
\value{

}
\description{
The equation for calculating similarity scores is as follows:
\deqn{exp(0.5 x (experimental.value - theoretical.value / flexibility)^2)}
}
\details{
Similarity scores between experimental and theoretical values are calculated using the above function.
The background hypothesis of the equations for accurate mass and retention time similarity is
that the differences between experimental and theoretical values follow a Gaussian distribution.
For more information, please refer to "MS-DIAL: Data Independing MS/MS Deconvolution for Comprehensive
Metabolome Analysis" Nat. Methods. 2015 June. doi: 10.1038/nmeth.3393.
}
