% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkContaminants}
\alias{checkContaminants}
\title{Check known contaminants}
\usage{
checkContaminants(experimental.df, common.contaminants, ppm.tolerance)
}
\arguments{
\item{experimental.df}{Experimental dataframe, containing at least the columns primary_key (numeric), MassFeature (character),
mz_experimental (numeric), and z_experimental (numeric).}

\item{common.contaminants}{A dataframe listing known contaminants for comparison. This csv can be found on the Shared Ingalls Google Drive
in the MARS_project folder.}

\item{ppm.tolerance}{Acceptable level of parts per million (ppm). Default is 15.}
}
\value{
matchedKnownContaminants, a small dataframe containing potential contaminants. These can be matched to the
annotated dataframe by the primary_key column.
}
\description{
Check known contaminants
}
