% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ConcatToScan}
\alias{ConcatToScan}
\title{ConcatToScan}
\usage{
ConcatToScan(concat.format)
}
\arguments{
\item{concat.format}{MS2 data in "concatenated format": mz and intensity separated by comma,
and separated from the next set of fragments by a semicolon. For example, "148.0, 100; 147.0, 50;", etc.}
}
\value{
scantable: The same MS2 data separated into a mini dataframe of mz and intensity.
}
\description{
ConcatToScan
}
