% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{IsolateMoNACandidates}
\alias{IsolateMoNACandidates}
\title{Compare experimental mass features to scraped MoNA data, based on MS1 and MS2 information.}
\usage{
IsolateMoNACandidates(MoNA.Mass, experimental.df, potential.candidates)
}
\arguments{
\item{MoNA.Mass}{Single MoNA mass, isolated from scraped MoNA df.}

\item{experimental.df}{experimental dataframe, islated to contain only MS1 and MS2 data.}
}
\value{
final.candidates: dataframe of experimental data, matched with information from MoNA.
}
\description{
Compare experimental mass features to scraped MoNA data, based on MS1 and MS2 information.
}
