% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MakeMS2CosineDataframe}
\alias{MakeMS2CosineDataframe}
\title{Create a filtered mini dataframe (a scantable) from two columns in a user-defined dataframe.}
\usage{
MakeMS2CosineDataframe(df)
}
\arguments{
\item{df}{A dataframe containing "scan1" and "scan2" columns.}
}
\value{

}
\description{
Create a filtered mini dataframe (a scantable) from two columns in a user-defined dataframe.
}
