% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeScantable.R
\name{MakeScantable}
\alias{MakeScantable}
\title{Create a filtered mini dataframe from a concatenated scanlist of MS2s.}
\usage{
MakeScantable(scan)
}
\arguments{
\item{scan}{Single observation from a dataframe containing MS2 m/z and intensity spectra, separated by semicolons.}
}
\value{
scantable: A tiny dataframe, containing columns of mz and intensity.
Intensity is scaled to 100 and filtered to drop all intensity values below 0.5.
}
\description{
Create a filtered mini dataframe from a concatenated scanlist of MS2s.
}
