% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatchConfidenceLevel1.R
\name{MatchConfidenceLevel1}
\alias{MatchConfidenceLevel1}
\title{MatchConfidenceLevel1}
\usage{
MatchConfidenceLevel1(
  experimental.values,
  theoretical.values,
  mz.flexibility,
  rt.flexibility
)
}
\arguments{
\item{experimental.values}{A dataframe of unknown experimental values, created by the user.
It is important to note that it must be in the following format, including capitalization and class:
mz (numeric), rt (numeric, in seconds), column (numeric), z (numeric), MS2 (in concatenated format, character).}

\item{theoretical.values}{TODO A dataframe of known, theoretical values, i.e, the Ingalls Standards list.
Like its partner the experimental.values dataframe, the theoretical.values dataframe columns need to be in
the following exact format: compound (character), mz (numeric), rt (numeric, in seconds),
column (all caps, character), z (numeric), OPTIONAL filename (character), and MS2 (concatenated format, character).}

\item{mz.flexibility}{TODO Usually defined as 0.02.}

\item{rt.flexibility}{Usually defined as ~ 15-30 seconds.}
}
\value{
A complete dataframe, annotated for Confidence Level 1.
}
\description{
Compare a dataframe of experimental values with a dataframe of all theoretical values from the Ingalls
Lab Standards sheet.
}
