% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{MS2CosineSimilarity}
\alias{MS2CosineSimilarity}
\title{Finds the weighted cosine similarity between two sets of MS2 spectra.}
\usage{
MS2CosineSimilarity(scan1, scan2)
}
\arguments{
\item{scan1}{Tiny dataframe of MS2, from first set of values to be compared. Column 1 is mz, column 2 is intensity. Dataframe is the output of the MakeScantable() function.}

\item{scan2}{Tiny dataframe of MS2, from second set of values to be compared. Column 1 is mz, column 2 is intensity. Dataframe is the output of the MakeScantable() function.}
}
\value{
cosine.similarity: A weighted similarity score between 0 and 1, indicating the cosine relationship of the two vectors.
}
\description{
Finds the weighted cosine similarity between two sets of MS2 spectra.
}
