% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pasteWithoutNA}
\alias{pasteWithoutNA}
\title{Paste items together, ignoring NAs}
\usage{
pasteWithoutNA(..., sep = "; ")
}
\arguments{
\item{...}{List of items to be pasted}

\item{sep}{Defined separator of ";" to remain consistent with MARS concatenation protocols}
}
\value{
Pasted item, ignoring the character "NA"s produced by regular paste.
}
\description{
Paste items together, ignoring NAs
}
