% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ReplaceNA}
\alias{ReplaceNA}
\title{Remove character values of "NA;" found in MARS outputs.}
\usage{
ReplaceNA(column)
}
\arguments{
\item{column}{Character column that contains one or more "NA;" values.}
}
\value{

}
\description{
Remove character values of "NA;" found in MARS outputs.
}
