% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ScanToConcat}
\alias{ScanToConcat}
\title{ScanToConcat}
\usage{
ScanToConcat(scantable.format)
}
\arguments{
\item{scantable.format}{MS2 data in scantable format: a mini dataframe of mz and intensity.}
}
\value{
concatenated: the same MS2 data converted to a concatenated format of paured mz and intensity
values separated by semicolons.
}
\description{
ScanToConcat
}
