% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateMoNAConfidenceLevel2.R
\name{AnnotateMoNAConfidenceLevel2}
\alias{AnnotateMoNAConfidenceLevel2}
\title{Annotate Confidence Level 2: MoNA}
\usage{
AnnotateMoNAConfidenceLevel2(
  Confidence.Level.1,
  MassBank.Neg,
  MassBank.Pos,
  mz.flexibility
)
}
\arguments{
\item{Confidence.Level.1}{A dataframe of experimental values, already annotated for Confidence Level 1.
Please refer to the AnnotateConfidenceLevel1() function for details!}

\item{MassBank.Neg}{Spectra in negative ion mode from MassBank of North America, scraped and downloaded. A modified example is included in this package,
but the complete csv is available on the Ingalls Shared Drive in the MARS_Project folder, titled NEG_Spectra.csv}

\item{MassBank.Pos}{Spectra in positive ion mode from MassBank of North America, scraped and downloaded. A modified example is included in this package,
but the complete csv is available on the Ingalls Shared Drive in the MARS_Project folder, titled POS_Spectra.csv}

\item{mz.flexibility}{Flexibility in Daltons for MS1 matching between experimental and theoretical values. Usually defined as 0.02.}
}
\value{
A dataframe annotated for Confidence Level 2, in addition to the Confidence Level 1 annotation from the previous step.
}
\description{
This function takes a dataframe of experimental values, already annotated for Confidence Level 1, and compares it to theoretical values
from MassBank of North America (MoNA). Those compounds that have a good match (see README documentation for specifics on what defines a "good match")
to MoNA MS1 & MS2 receive a Confidence Level 2 ranking.
In order to run this code, external data from MoNA is required.
TODO The code to produce and/or update the scraped spectra from MassBank is contained in the TODOFUNCTION as part of the phobos package.
}
\examples{
library(phobos)
# Load experimental data that has already been annotated for Confidence Level 1 using the AnnotateConfidenceLevel1 function.
# Remember that dataframe format is important: please refer to README or the documentation for AnnotateConfidenceLevel1() for details.

Confidence.Level.1 <- read.csv("data/Example_ConfidenceLevel1.csv")
MassBank.Neg <- read.csv("data/NEG_Spectra.csv")
MassBank.Pos <- read.csv("data/POS_Spectra.csv")
Example_ConfidenceLevel2 <- AnnotateMoNAConfidenceLevel2(Confidence.Level.1 = Confidence.Level.1, MassBank.Neg = MassBank.Neg,
MassBank.Pos = MassBank.Pos, mz.flexibility = 0.02

}
