% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateMoNAConfidenceLevel2.R
\name{AnnotateMoNAConfidenceLevel2}
\alias{AnnotateMoNAConfidenceLevel2}
\title{Compare experimental values to MassBank of North America (MoNA) in order to annotate unknown values for confidence level 2: a good match to MS1 and MS2.}
\usage{
AnnotateMoNAConfidenceLevel2(
  Confidence.Level.1,
  mz.flexibility,
  rt.flexibility
)
}
\arguments{
\item{Confidence.Level.1}{A dataframe of experimental values, already annotated for Confidence Level 1.
Please refer to the AnnotateConfidenceLevel1 function for details!}

\item{mz.flexibility}{Flexibility for m/z matching between experimental and theoretical values. Usually defined as 0.02.}

\item{rt.flexibility}{Flexibility for retention time matching between experimental and theoretical values. Usually defined as ~ 15-30 seconds.}

\item{z}{Choose positive or negative charge for appropriate value comparison. The options are 1 for positive mode and -1 for negative.}
}
\value{
A complete dataframe, annotated for Confidence Level 1.
}
\description{
Compare experimental values to MassBank of North America (MoNA) in order to annotate unknown values for confidence level 2: a good match to MS1 and MS2.
}
\examples{
#' library(phobos)
# Load experimental data that has already been annotated for Confidence Level 1 using the AnnotateConfidenceLevel1 function.
# Remember that dataframe format is important: please refer to README or the documentation for AnnotateConfidenceLevel1() for details.

example_dir <- system.file("example_data", package = "phobos")
example_data <- list.files(example_dir, full.names = TRUE)
Confidence.Level.1 <- read.csv(grep("Example_ConfidenceLevel1", example_data, value = TRUE))
z <- 1
example_confidenceL2 <- AnnotateMoNAConfidenceLevel2(Confidence.Level.1 = Confidence.Level.1, z = z,
mz.flexibility = 0.02, rt.flexibility = 30)
}
