% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AnnotateMoNAConfidenceLevel3.R
\name{AnnotateMoNAConfidenceLevel3}
\alias{AnnotateMoNAConfidenceLevel3}
\title{Annotate Confidence Level 3: MoNA}
\usage{
AnnotateMoNAConfidenceLevel3(Confidence.Level.2, mz.flexibility)
}
\arguments{
\item{Confidence.Level.2}{A dataframe of experimental values, already annotated for Confidence Levels 1 & 2.
Please refer to the README or documentation for the AnnotateConfidenceLevel1 and AnnotateMoNAConfidenceLevel2 functions for details.}

\item{mz.flexibility}{Flexibility for m/z matching between experimental and theoretical values. Usually defined as 0.02.}
}
\value{
A complete dataframe, annotated for Confidence Level 3 using MoNA.
}
\description{
Annotate Confidence Level 3: MoNA
}
\examples{
library(phobos)
# Load experimental data that has already been annotated for Confidence Levels 1 & 2
# using the AnnotateConfidenceLevel1 and AnnotateMoNAConfidenceLevel2 function2.
# Remember that dataframe format is important: please refer to README or the documentation for AnnotateConfidenceLevel1() for details.

example_dir <- system.file("example_data", package = "phobos")
example_data <- list.files(example_dir, full.names = TRUE)
Confidence.Level.2 <- read.csv(grep("Example_ConfidenceLevel2", example_data, value = TRUE))

example_confidenceL3 <- AnnotateMoNaConfidenceLevel3(Confidence.Level.2 = Confidence.Level.2, mz.flexibility = 0.02)
}
